requirements.txt
streamlit==1.36.0
python-dotenv==1.0.1
openai==1.35.10
Pillow==10.4.0
pytz==2024.1